Refactor SubjectCard with StaticMedia and InteractiveMedia components
