+++ packages/lib-react-components/src/SubjectCard/SubjectCard.jsx
@@ -1,14 +1,15 @@
 import { Anchor, Box } from 'grommet'
-import mime from 'mime/lite'
 import { arrayOf, bool, node, object, objectOf, oneOf, shape, string } from 'prop-types'
 import styled from 'styled-components'
 
 import { useTranslation } from '../translations/i18n'
-import Media from '../Media'
 import MetadataIconButton from '../MetadataIconButton'
 import FavoritesIconButton from '../FavoritesIconButton'
 import CollectIconButton from '../CollectIconButton'
 import ShareIconButton from '../ShareIconButton'
+
+import StaticMedia from './components/StaticMedia'
+import InteractiveMedia from './components/InteractiveMedia'
 
 const METATOOLS_HEIGHT = 45
 
@@ -16,7 +17,7 @@ const StyledSubjectCard = styled(Box)`
   position: relative;
 `
 
-const StyledImageLink = styled(Anchor)`
+const StyledMediaLink = styled(Anchor)`
   display: block;
   text-decoration: none;
   width: 100%;
@@ -24,43 +25,6 @@ const StyledImageLink = styled(Anchor)`
   &:hover {
     text-decoration: none;
   }
-`
-
-const StyledPreview = styled(Box)`
-  overflow: hidden;
-  position: relative;
-`
-
-const StyledBackground = styled(Box)`
-  filter: blur(12px);
-  inset: 0;
-  position: absolute;
-  transform: scale(1.2); // scale up the background to hide edges created by the blur
-  z-index: 0;
-`
-
-const StyledForegroundMedia = styled(Media)`
-  position: relative;
-  z-index: 1;
-`
-
-const StyledTitle = styled(Box)`
-  background: linear-gradient(180deg, rgba(0, 0, 0, 0) 0%, rgba(0, 0, 0, 0.7) 100%);
-  bottom: 0;
-  height: 72px;
-  left: 0;
-  position: absolute;
-  right: 0;
-  text-align: center;
-  z-index: 2;
-`
-
-const StyledTitleText = styled.span`
-  color: ${props => props.theme.global.colors.white};
-  font-size: 1rem;
-  font-weight: 400;
-  letter-spacing: 0.8px;
-  text-shadow: 0 1px 2px rgba(0, 0, 0, 0.6);
 `
 
 function cardWidth(size) {
@@ -116,11 +80,7 @@ function SubjectCard({
   const metaToolsSectionGap = metaToolsGap(size)
 
   // subject properties
-  const { locations, metadata } = subject
-  const mediaSrc = locations?.[0] ? Object.values(locations[0])[0] : null
-  const mimeType = mediaSrc ? mime.getType(mediaSrc) : null
-  const [ mediaType ] = mimeType ? mimeType.split('/') : []
-  const showBackground = mediaType === 'image' || mediaType === 'video'
+  const { metadata } = subject
   const subjectTalkHref = (projectSlug && subject?.id) ? `/projects/${projectSlug}/talk/subjects/${subject.id}` : undefined
   const origin = typeof window !== 'undefined' ? window.location.origin : ''
   const subjectTalkUrl = subjectTalkHref ? `${origin}${subjectTalkHref}` : undefined
@@ -137,52 +97,23 @@ function SubjectCard({
         a11yTitle={linkTitle}
         href={subjectTalkHref}
       >
-        <StyledPreview
-          height={`${previewHeight}px`}
-          round={{ corner: 'top', size: '8px' }}
-          width={`${width}px`}
-        >
-          {(mediaSrc && showBackground) ? (
-            <StyledBackground>
-              <Media
-                alt=''
-                controls={false}
-                fit='cover'
-                height={previewHeight}
-                src={mediaSrc}
-                width={width}
-                aria-hidden='true'
-                tabIndex={-1}
-              />
-            </StyledBackground>
-          ) : null}
-
-          {mediaSrc ? (
-            <StyledForegroundMedia
-              alt={subjectIdTitle}
-              controls={false}
-              fit='contain'
-              placeholder={placeholder}
-              height={previewHeight}
-              src={mediaSrc}
-              width={width}
-            />
-          ) : null}
-
-          {interactive ? null : (
-            <StyledTitle
-              align='center'
-              direction='row'
-              gap='xsmall'
-              justify='center'
-              pad={{ horizontal: 'small', vertical: 'medium' }}
-            >
-              <StyledTitleText>
-                {subjectIdTitle}
-              </StyledTitleText>
-            </StyledTitle>
-          )}
-        </StyledPreview>
+        {interactive ? (
+          <InteractiveMedia
+            placeholder={placeholder}
+            previewHeight={previewHeight}
+            subject={subject}
+            subjectIdTitle={subjectIdTitle}
+            width={width}
+          />
+        ) : (
+          <StaticMedia
+            placeholder={placeholder}
+            previewHeight={previewHeight}
+            subject={subject}
+            subjectIdTitle={subjectIdTitle}
+            width={width}
+          />
+        )}
       </StyledMediaLink>
       <Box
         align='center'
