+++ packages/lib-react-components/src/SubjectCard/components/InteractiveMedia/InteractiveMedia.jsx
@@ -1,0 +1,5 @@
+function InteractiveMedia() {
+  return null
+}
+
+export default InteractiveMedia
