+++ packages/lib-react-components/src/SubjectCard/components/InteractiveMedia/InteractiveMediaContainer.jsx
@@ -1,0 +1,7 @@
+import InteractiveMedia from './InteractiveMedia'
+
+function InteractiveMediaContainer() {
+  return <InteractiveMedia />
+}
+
+export default InteractiveMediaContainer
