+++ packages/lib-react-components/src/SubjectCard/components/InteractiveMedia/index.js
@@ -1,0 +1,1 @@
+export { default } from './InteractiveMediaContainer'
