+++ packages/lib-react-components/src/SubjectCard/components/StaticMedia/index.js
@@ -1,0 +1,1 @@
+export { default } from './StaticMediaContainer'
